Move workshop files into workshop/list folder
