+++ src/components/workshop/list/WorkshopCategoryIcon.tsx
